feat!: reconcile SDK resources in-process and drop Terraform CLI from image
Move all 106 Terraform Plugin SDK resources from upjet's CLI-based
workspace mode to the in-process SDKv2 architecture
(WithTerraformPluginSDKIncludeList + WithTerraformProvider). All
resources now reconcile natively against the compiled-in provider, so
the container image no longer ships the Terraform CLI (1.8.1, flagged
by security scans and BUSL-licensed) or the mirrored native provider
binary — only the provider Go binary remains.

- config: pass ovh.Provider() for in-process SDKv2 reconciliation and
  the (empty) CLIReconciledResourceList to override upjet's catch-all
  default include list
- controllers: regenerate; 196 controllers switch from
  tjcontroller.NewConnector to NewTerraformPluginSDKAsyncConnector
- image: remove terraform zip, provider filesystem mirror,
  terraformrc.hcl and related build args; bump base image from
  alpine 3.17.1 (EOL) to alpine 3.22
- docs: update AGENTS.md, README and the upgrade skill; the Terraform
  CLI is still used at build time to generate config/schema.json but
  never ships in the image

BREAKING CHANGE: CRD schemas are now generated from the compiled-in Go
provider schema instead of the JSON schema export. 506 numeric fields
across 86 CRDs change from type: number to type: integer
(format: int64), and spec.forProvider.projectId on the vrack
Cloudproject kind is now enforced as required.

Closes #64

+++ apis/cluster/cloud/v1alpha1/zz_plan_types.go
@@ -40,7 +40,7 @@ type PlanInitParameters_2 struct {
 
 	// Size of the Savings Plan
 	// Size of the Savings Plan
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 }
 
 type PlanObservation_2 struct {
@@ -85,7 +85,7 @@ type PlanObservation_2 struct {
 
 	// Billing ID of the service
 	// ID of the service
-	ServiceID *float64 `json:"serviceId,omitempty" tf:"service_id,omitempty"`
+	ServiceID *int64 `json:"serviceId,omitempty" tf:"service_id,omitempty"`
 
 	// ID of the public cloud project. If omitted, the OVH_CLOUD_PROJECT_SERVICE environment variable is used.
 	// ID of the public cloud project
@@ -93,7 +93,7 @@ type PlanObservation_2 struct {
 
 	// Size of the Savings Plan
 	// Size of the Savings Plan
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 
 	// Start date of the Savings Plan
 	// Start date of the Savings Plan
@@ -138,7 +138,7 @@ type PlanParameters_2 struct {
 	// Size of the Savings Plan
 	// Size of the Savings Plan
 	// +kubebuilder:validation:Optional
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 }
 
 // PlanSpec defines the desired state of Plan
