+++ apis/cluster/dns/v1alpha1/zz_zone_types.go
@@ -19,14 +19,18 @@ type OrderDetailsInitParameters struct {
 type OrderDetailsObservation struct {
 
 	// description
+	// description
 	Description *string `json:"description,omitempty" tf:"description,omitempty"`
 
+	// expiration date
 	// expiration date
 	Domain *string `json:"domain,omitempty" tf:"domain,omitempty"`
 
 	// order detail id
-	OrderDetailID *float64 `json:"orderDetailId,omitempty" tf:"order_detail_id,omitempty"`
+	// order detail id
+	OrderDetailID *int64 `json:"orderDetailId,omitempty" tf:"order_detail_id,omitempty"`
 
+	// quantity
 	// quantity
 	Quantity *string `json:"quantity,omitempty" tf:"quantity,omitempty"`
 }
@@ -157,7 +161,7 @@ type ZoneOrderObservation struct {
 
 	// order id
 	// order id
-	OrderID *float64 `json:"orderId,omitempty" tf:"order_id,omitempty"`
+	OrderID *int64 `json:"orderId,omitempty" tf:"order_id,omitempty"`
 }
 
 type ZoneOrderParameters struct {
